source/requirement.txt
certifi==2019.11.28
chardet==3.0.4
Django==1.11.23
django-cors-headers==3.2.0
djangorestframework==3.10.3
idna==2.8
pytz==2019.3
requests==2.22.0
urllib3==1.25.7
xlrd==1.2.0
XlsxWriter==1.2.7
